refactor: rename dir to form-template
