+++ assets/src/js/frontend/give.js
@@ -7,7 +7,7 @@ import 'iframe-resizer';
 
 // Give plugins.
 import '../plugins/give-hint.css';
-import { initializeIframeResize } from '../plugins/form-theme/utils';
+import { initializeIframeResize } from '../plugins/form-template/utils';
 
 // Give API.
 import GiveAPI from '../plugins/give-api/api';
@@ -17,8 +17,8 @@ import './give-donations';
 import './give-ajax';
 import './give-misc';
 import './give-donor-wall';
-import iFrameResizer from '../plugins/form-theme/iframe-content';
-import '../plugins/form-theme/parent-page';
+import iFrameResizer from '../plugins/form-template/iframe-content';
+import '../plugins/form-template/parent-page';
 
 const { init, fn, form, notice, cache, donor, util } = GiveAPI;
 window.Give = { init, fn, form, notice, cache, donor, util, initializeIframeResize };
